Int16 fixes

+++ lang/src/Parser/AST.hs
@@ -6,8 +6,8 @@ import Data.Int      (Int16)
 
 type Module = [Expr]
 
-data Expr = Lit Integer
-          | VecLit [Integer]
+data Expr = Lit Int16
+          | VecLit [Int16]
           | VarE VarName
           | BinOp Op Expr Expr
           | If Expr [Expr] [Expr]
@@ -17,7 +17,7 @@ data Expr = Lit Integer
 
 
 type VarName = String
-data Type    = Scalar | Vector Integer deriving (Show, Eq, Ord)
+data Type    = Scalar | Vector Int16 deriving (Show, Eq, Ord)
 
 
 data Op = Add Type | Sub Type | Mul Type | Div Type deriving (Show, Eq, Ord)
